yyonline/requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.9.8
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==0.5.1
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
jdcal==1.3
oauthlib==2.0.6
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
Pillow==4.2.0
PyJWT==1.5.3
PyMySQL==0.7.11
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.1.0
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-core==1.5.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
userprofile==0.2.5
xlrd==1.1.0
xlwt==1.3.0
